Use HeaderMap from http crate for ResponseDefinition
The http crate was already transitively included
and provides a high quality HeaderMap implementation.
This map implementation takes care of case insensitive header name matching
and allows multiple values for a single key.

A custom serialization and deserialization implementation was written
to conditionally serialize and serialize a single string value
or a sequence of strings values depending on the values count.